number of scheduled trips for route and agency metric
